refactor: read and write
